drApi/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
lazy-object-proxy==1.10.0
Markdown==3.6
pillow==10.4.0
pytz==2024.1
redis==5.0.7
sqlparse==0.5.0
typing_extensions==4.12.2
validate-docbr==1.10.0
